LPS-62127 Make css mime type logic accesible to other portlets
