+++ modules/apps/document-library/document-library-api/src/main/java/com/liferay/document/library/configuration/DLConfiguration.java
@@ -28,6 +28,16 @@ import com.liferay.configuration.admin.ConfigurationAdmin;
 )
 public interface DLConfiguration {
 
+	@Meta.AD(
+		deflt = "text/asp|text/css|text/ecmascript|text/html|text/javascript|text/x-c|text/x-fortran|text/x-java-source|text/x-pascal|text/x-script.perl|text/x-script.perl-module|text/xml", required = false
+	)
+	public String[] codeFileMimeTypes();
+
+	@Meta.AD(
+		deflt = "application/x-7z-compressed|application/x-ace-compressed|application/x-compressed|application/x-rar-compressed|application/x-zip-compressed|application/zip", required = false
+	)
+	public String[] compressedFileMimeTypes();
+
 	/**
 	 * Set the location of the XML file containing the configuration of the
 	 * default display templates for the Document Library portlet.
@@ -38,6 +48,19 @@ public interface DLConfiguration {
 	)
 	public String displayTemplatesConfig();
 
+	@Meta.AD(deflt = "audio|image|video", required = false)
+	public String[] multimediaFileMimeTypes();
+
+	@Meta.AD(
+		deflt = "application/mspowerpoint|application/powerpoint|application/vnd.apple.keynote|application/vnd.ms-powerpoint|application/vnd.oasis.opendocument.presentation|application/vnd.openxmlformats-officedocument.presentationml.presentation|application/x-mspowerpoint", required = false
+	)
+	public String[] presentationFileMimeTypes();
+
+	@Meta.AD(
+		deflt = "application/excel|application/vnd.ms-excel|application/vnd.oasis.opendocument.spreadsheet|application/vnd.openxmlformats-officedocument.spreadsheetml.sheet|application/vnd.sun.xml.calc|application/x-excel|application/x-msexcel", required = false
+	)
+	public String[] spreadSheetFileMimeTypes();
+
 	/**
 	 * Set the interval in hours on how often
 	 * TemporaryFileEntriesMessageListener will run to check for expired
@@ -45,5 +68,13 @@ public interface DLConfiguration {
 	 */
 	@Meta.AD(deflt = "1", required = false)
 	public int temporaryFileEntriesCheckInterval();
+
+	@Meta.AD(
+		deflt = "application/msword|application/vnd.oasis.opendocument.text|application/vnd.openxmlformats-officedocument.wordprocessingml.document|text/plain", required = false
+	)
+	public String[] textFileMimeTypes();
+
+	@Meta.AD(deflt = "application/pdf", required = false)
+	public String[] vectorialFileMimeTypes();
 
 }
