+++ modules/apps/wiki/wiki-web/build.gradle
@@ -3,6 +3,7 @@ dependencies {
 	compile group: "org.apache.struts", name: "struts-core", version: "1.3.10"
 	compile project(":apps:application-list:application-list-api")
 	compile project(":apps:configuration-admin:configuration-admin-api")
+	compile project(":apps:document-library:document-library-api")
 	compile project(":apps:dynamic-data-mapping:dynamic-data-mapping-taglib")
 	compile project(":apps:item-selector:item-selector-api")
 	compile project(":apps:item-selector:item-selector-criteria-api")
